Fixed namespace names after code reorganization
