+++ aspnetcore-webapi/UsingResultPatternInNETWebAPI/Tests/GlobalUsings.cs
@@ -6,10 +6,3 @@ global using System.Net.Http.Json;
 global using UsingResultPatternInNETWebAPI;
 global using UsingResultPatternInNETWebAPI.Dtos;
 global using UsingResultPatternInNETWebAPI.Repositories;
-global using UsingResultPatternInNETWebAPI.v1.Services;
-global using UsingResultPatternInNETWebAPI.v2.Services;
-global using UsingResultPatternInNETWebAPI.v3.Exceptions;
-global using UsingResultPatternInNETWebAPI.v3.Services;
-global using UsingResultPatternInNETWebAPI.v4.Services;
-global using UsingResultPatternInNETWebAPI.v5.Errors;
-global using UsingResultPatternInNETWebAPI.v5.Services;
