+++ aspnetcore-webapi/UsingResultPatternInNETWebAPI/Tests/Services/ExceptionsForFlowControlContactServiceIntegrationTest.cs
@@ -1,4 +1,7 @@
-﻿namespace Tests;
+﻿using UsingResultPatternInNETWebAPI.Exceptions;
+using UsingResultPatternInNETWebAPI.Services;
+
+namespace Tests.Services;
 
 public class ExceptionsForFlowControlContactServiceTests
 {
