+++ aspnetcore-webapi/UsingResultPatternInNETWebAPI/Tests/Services/FluentResultsContactServiceIntegrationTest.cs
@@ -1,4 +1,7 @@
-﻿namespace Tests;
+﻿using UsingResultPatternInNETWebAPI.Errors;
+using UsingResultPatternInNETWebAPI.Services;
+
+namespace Tests.Services;
 
 public class FluentResultsContactServiceTests
 {
