+++ aspnetcore-webapi/UsingResultPatternInNETWebAPI/UsingResultPatternInNETWebAPI/Controllers/v1/BasicContactController.cs
@@ -1,4 +1,6 @@
-﻿namespace UsingResultPatternInNETWebAPI.v1.Controllers;
+﻿using UsingResultPatternInNETWebAPI.Services;
+
+namespace UsingResultPatternInNETWebAPI.Controllers.v1;
 
 [ApiController]
 [Route("api/v1/contacts")]
