+++ aspnetcore-webapi/UsingResultPatternInNETWebAPI/UsingResultPatternInNETWebAPI/Exceptions/DefaultExceptionHandler.cs
@@ -1,6 +1,6 @@
 ﻿using System.Net;
 
-namespace UsingResultPatternInNETWebAPI.v3.Exceptions;
+namespace UsingResultPatternInNETWebAPI.Exceptions;
 
 public class DefaultExceptionHandler : IExceptionHandler
 {
