+++ aspnetcore-webapi/UsingResultPatternInNETWebAPI/UsingResultPatternInNETWebAPI/Exceptions/ValidationException.cs
@@ -1,4 +1,4 @@
-﻿namespace UsingResultPatternInNETWebAPI.v3.Exceptions;
+﻿namespace UsingResultPatternInNETWebAPI.Exceptions;
 
 public class ValidationException : Exception
 {
