+++ aspnetcore-webapi/UsingResultPatternInNETWebAPI/UsingResultPatternInNETWebAPI/TheResultPattern/CustomError.cs
@@ -1,4 +1,4 @@
-﻿namespace UsingResultPatternInNETWebAPI.v4.TheResultPattern;
+﻿namespace UsingResultPatternInNETWebAPI.TheResultPattern;
 
 public sealed record CustomError(string Code, string Message)
 {
